+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes in this release:
 - Fixed plugin loading compatibility with compiler's import mechanism (#46, #49)
 - Added `inmanta_plugins` fixture to make abstraction of required module reloading when the compiler project is reset (related to #49)
 - Added deprecation warning for `project_no_plugins` fixture in favor of `INMANTA_TEST_NO_LOAD_PLUGINS` environment variable (#66)
+- Added resource unittest::IgnoreResource.
 
 # V 1.3.0
 Changes in this release: